[asyncmysql] Set VIO blocking with given status
Summary:
When we fetch a row, the VIO goes to blocking state for reasons that I
don't fully understand. The problem is in vio_read it's check if the vio
is blocking, if it is, then it sets to non-blocking. That's the normal
flow for non-SSL connections.

When in Multi Query, the state isn't being reset to nonblocking because the
ssl_set_blocking doesn't reset the value vio read checks.

It's not clear to me why this is done like this. So I want to see tests run.

Test Plan:
Built in tp2 and this fixed the issue where in a MultiQuery in SSL the "next_result" is being done blocking.

Reviewers: anca

Subscribers: