Fix several rocksdb_rpl tests to run more consistently
Summary:
The rocksdb_rpl tests issue kill signals to the mysqld server. But the
expect file needs to be created in order for the mtr framework to detect
the crash as expected. Otherwise, the mtr framework can fail the test.

Squash with https://reviews.facebook.net/D31365

Test Plan: run the tests

Reviewers: jtolmer, spetrunia, jkedgar, yoshinorim

Reviewed By: yoshinorim

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D51837