Issue #110: SQL command checksum returns inconsistent result
Summary:
When unpacking NULL values, we set the null-bit of the field to one.
Besides that, also set the field value to the default value (like
InnoDB does).
CHECKSUM TABLE depends on this as it includes field *value* in checksum
calculations even when the field has NULL value (this is despite the fact
that NULL-bits are also included in checksum calculation).

Test Plan: rocksdb.checksum_table test

Reviewers: yoshinorim, jtolmer, jkedgar, hermanlee4

Reviewed By: hermanlee4

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D51525