Fix checking for RocksDB recovery line in rpl_gtid_crash_safe_wal_corrupt
Summary:
The test moves the mysqld.2.err log to a new location during each test
run, but it would be better to maintain the entire history of the log
for debugging and archival purposes. Otherwise, when the test fails, the
initial report shows there is nothing in the log and can be misleading.

Squash with https://reviews.facebook.net/D43947

Test Plan: Run the unit test.

Reviewers: jtolmer, spetrunia, yoshinorim, jkedgar

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D51861