Disallow running rocksdb.deadlock in valgrind
Summary:
The rocksdb.deadlock test was failing because the default 1 second lock wait timeout was too low for running under valgrind. The test passes after increasing the lock wait timeout, but it takes 40 minutes to run. It might be better to just not run this test under valgrind.

Squash with: https://reviews.facebook.net/D48669

Test Plan:
./mtr --valgrind rocksdb.deadlock

Test is skipped

Reviewers: jkedgar hermanlee4

Subscribers: