Disallow running rocksdb.deadlock in valgrind and increase lock wait timeout
Summary:
The rocksdb.deadlock test was failing because the default 1 second lock wait timeout was too low. Increase this to 60 seconds.

This test takes 40 minutes to run under valgrind, so it might be better to just not run this test under valgrind.

Squash with: https://reviews.facebook.net/D48669

Test Plan:
./mtr --valgrind rocksdb.deadlock

Test is skipped

Reviewers: jkedgar, hermanlee4

Reviewed By: hermanlee4

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D51993