Forward connection certificate to audit plugin
Summary:
Forward the connection certificate to the audit plugin. The connection certificate can then be parsed by the audit plugin and handled appropriately.

This is done by caching a BUF_MEM struct on the THD object. Since it's not possible to change certificates on the same connection, this caching should be correct. The BUF_MEM is released on THD::release_resources.

This also fixed a memory leak in get_peer_cert_info where a call to X509_free was missing.

Test Plan:
  mysqltest.sh --valgrind main.information_schema_authinfo

  mysqltest.sh --manual-gdb main.information_schema_authinfo
Then check to see if certificate information is present in gdb.

Reviewers: jkedgar hermanlee4

Subscribers: jtolmer anca