Disable background compaction for type*index tests
Summary:
These tests depend on statistics to generate their query plans. With
drop index compactions occurring, if the memtable is flushed out, it is
possible for the statistics to change and affect the query plan. For
these tests, disabling background compaction will keep the tables all
within the memtable.

The downside with disabling compaction is the test server could stall
once all memtables are filled up, but given the size of the memtables
and the amount of data inserted by each test, the test should not hit
those limits.

Squash with https://reviews.facebook.net/D49107

Test Plan: Run unit tests

Reviewers: spetrunia, yoshinorim, jkedgar, jtolmer, rven

Reviewed By: rven

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D52395