Add rocksdb_stress test suite with a new load_generator
Summary:
Add a new load_generator (based on the innodb_stress) one that can
persist through server failures and reconnect after the server comes up.
The load_generator maintains a map of rows that have been
modified/removed and a list of rows of older transactions that were
updated. It checks the status of the current pending transaction and
verifies it was either committed or rolled back.

Test Plan: run the unit test

Reviewers: jtolmer, yoshinorim, jkedgar, santoshb

Subscribers: