Fix non-deterministic test failure
Summary: The main.slow_log_extra_big test had a non-deterministic failure.  The logging for some of the cleanup occurred earlier in the log than the last line that we intentionally logged.  This fix redirects the logging back to the original log file after the test is finished but before the cleanup occurs to attempt to avoid the problem.

Test Plan: MTR

Reviewers: hermanlee4, yoshinorim

Subscribers: