Forward connection certificate to audit plugin
Summary:
Forward the connection certificate to the audit plugin. The connection certificate can then be parsed by the audit plugin and handled appropriately. It made more sense for the certificate to live in the connection events, since they generally don't change between every general event, so the move was done.

This is done by caching a BUF_MEM struct on the THD object. Since it's not possible to change certificates on the same connection, this caching should be correct. The BUF_MEM is released on THD::release_resources.

If upstream bumps the MYSQL_AUDIT_INTERFACE_VERSION, we should bump ours to be greater or equal to it.

Test Plan: Test with plugin to see that certificate information is present.

Reviewers: jtolmer, jkedgar

Reviewed By: jkedgar

Subscribers: webscalesql-eng, anca

Differential Revision: https://reviews.facebook.net/D52263