Forward connection certificate to audit plugin
Summary:
Forward the connection certificate to the audit plugin. The connection certificate can then be parsed by the audit plugin and handled appropriately.

This is done by caching a BUF_MEM struct on the THD object. Since it's not possible to change certificates on the same connection, this caching should be correct. The BUF_MEM is released on THD::release_resources.

Test Plan: Test with plugin to see that certificate information is present.

Reviewers: jkedgar

Subscribers: webscalesql-eng, anca

Differential Revision: https://reviews.facebook.net/D52263