Fix myrocks hang when dropping a database
Summary:
When a new transaction is created during ha_rocksdb::start_stmt due to
autocommit being disabled, it needs to be registered with mysql.
Otherwise, it does not get cleaned up correctly and can lead to the
server crashing or hanging.

Test Plan: run unit tests and add a new test case

Reviewers: spetrunia, yoshinorim, jkedgar, jtolmer

Subscribers: