Stabilize the slow_log_extra and slow_log_extra_big tests
Summary:
Recently the main.slow_log_extra_big test failed and while looking into it I discovered that the test will fail quickly when crushed and that main.slow_log_extra also has the same problem.  It appears that the problem is that the slow log output lines are sent to the slow log asynchronously which means that they may not be there when the test parses the log file for them.  This diffs adds a method to wait for N lines in a file matching a pattern to exist in a file so that we can wait until the lines exist before parsing the log file.

Squash with: cc43636 Stabilize main.slow_log_extra test
Note that this supercedes D52473 which got merged with   9dec869 Port per query stats from 5.1 to 5.6
This may cause some merge conflicts during a rebase.

Test Plan: MTR and crushing main.slow_log_extra and main.slow_log_extra_big

Reviewers: hermanlee4, jtolmer

Subscribers: