Fix asan failure in rocksdb.optimize_table due to new myrocks logger
Summary:
During some test runs of rocksdb.optimize_table, there would be a
strange ASan failure when the server is shutdown in the middle of the
test. It is not clear what the root cause of the failure is, but it
looks to be triggered by myrocks setting the log_level_ member of the
rocksdb::Logger class through the SetInfoLogLevel() function. Setting
the variable through the constructor also causes the obscure ASan
problem.

This diff replicates the log_level member behavior, but in the RdbLogger
class, allowing myrocks to ignore the log_level_ member. The base
class's Logv routines are not called, so not having the level set
correctly is fine.

Also rename Logger to RdbLogger to help differentiate it from
rocksdb::Logger.

Squash with https://reviews.facebook.net/D52419

Test Plan: run unit tests

Reviewers: spetrunia, yoshinorim, jkedgar, jtolmer

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D52731