Fixed the issue that mysqld creates 32K threads
Summary:
We have seen that mysqld creates much more threads than @@max_connections. The bug was that when a thread is finishing, the connection count is decremented too early, it takes a little time for the thread to clean up and complete exit but since the connection count has been decremented so new thread will be created for the new connections.

In this fix, LOCK_thread_count instead of LOCK_connection_count is used to protect connection_count. When a thread finishes, it will either be blocked for reuse, or exit. When it needs to be blocked, connection_count will get decremented while holding the LOCK_thread_count before waiting on COND_thread_cache so connection_count is very well synchronized. But when the thread needs to exit, connection_count will get decremented right before pthread_exit() call, which is the best can be done, but the thread won't be complete gone yet even after pthread_exit() and still visible by "ps -eLf " (usually pthread_join() is needed to wait for the thread completely gone but this is not suitable for mysql threading model), and since connection_count has been decremented already so thread may be created for new connections, this will make the actual peak thread number a little bit more than max_connections in very short period of time but this won't be a problem.

After this fix, the total number of threads for connections in mysqld is very well caped by @@max_connections. I also see the overall throughput in terms of max number of connections per second gets improved 1-2% that I believe attribute to less overhead of too much threads piling up in mysqld.

Test Plan: MTR and Performace tests

Reviewers: jtolmer, tianx

Subscribers: webscalesql-eng