Add lock information to lock timeout error message
Summary:
Currently, on lock timeout, we only give
  ERROR 1205 (HY000): Lock wait timeout exceeded; try restarting transaction
 which has no information on what type of lock timed out.

This change augments the message with the lock type to say something like
  ERROR 1205 (HY000): Lock wait timeout exceeded; try restarting transaction: Timeout on index: test/t1.GEN_CLUST_INDEX
which shows which type of lock timed out as well as the associated object.

There two cases to handle here. For error messages coming from the storage engine, the error is stored onto `trx_t::detailed_error`, which will get forwarded to the SQL layer for display. For MDL locks in the SQL layer, we are directly using the information in the MDL lock to display the error message.

Note that this isn't implemented for RocksDB, only for InnoDB.

Test Plan: mtr

Reviewers: jtolmer

Subscribers: ebergen, webscalesql-eng

Differential Revision: https://reviews.facebook.net/D47613