TRUNCATE TABLE to reclaim space in MyRocks (#137)
Summary:
When truncating a table, space was not really reclaimed in MyRocks.
This was because ha_rocksdb::delete_table() was not called via truncate.
When using HTON_CAN_RECREATE flag, ha_rocksdb::create() should check
if table already exists or not, and delete indexes if exists. This diff
follows the logic.

Differential Revision: https://reviews.facebook.net/D53031