MyRocks: Returning an error on ROLLBACK TO SAVEPOINT if modifying rows (#52)
Summary:
MyRocks did neither rollback nor return an error on ROLLBACK TO
SAVEPOINT, which was confusing to users. This diff makes MyRocks return
an error on ROLLBACK TO SAVEPOINT if modifying rows. Users can
at least know that.
Real fix is to support rollback on savepoint, not to return an error,
which will happen in future release.

Differential Revision: https://reviews.facebook.net/D55161

fbshipit-source-id: 228f0131f26