Use PIC versions of cryptography and OpenSSL libraries if present
Summary: In a situation where we have both PIC and non-PIC versions of various libraries present in most cases we'll use a non-PIC version. However, for some of the MySQL shared libraries which utilize cryptography and OpenSSL, PIC version is needed. Therefore change the `find-library()` search order (most-specific to least-specific for common use cases) to first look at PIC versions of the subset of libraries.

Test Plan:

- Build in the environment where libraries haven't been separated into PIC and non-PIC versions and observe the build to succeed.
- Build in the environment where libraries have been separated into PIC and non-PIC versions:

  1) Observe a build break.
  2) Make the changes to `ssl.cmake` and notice that this causes the build to succeed.

Reviewers: jtolmer

Reviewed By: jtolmer

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D53019