Fix performance of reporting snapshot status
Summary:
The original diff to supply the ability to report snapshot statuses used a global table and acquired a mutex to add or remove a snapshot from this table.  This has proved to be a performance problem.  Update the code to keep the snapshot information per Rdb_transaction and have a way to walk a list of Rdb_transactions to query this information

Squash with: https://reviews.facebook.net/D45945

Test Plan: MTR

Reviewers: hermanlee4, yoshinorim

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D53463