Minor efficiency gains while updating statistics
Summary:
I noticed that while gathering the statistics that the code to handle the sliding window of deletes was not optimal.  It always cleared the deleted flag before then checking to see if it would set it again.  If the current flag and the new flag are the same we can skip everything else.  In addition it always updated the maximum deleted rows value, but that check only needs to happen when we increment the number of deleted rows - it is useless when decrementing the number of deleted rows.

Squash with: https://reviews.facebook.net/D52749
These two may also be able to be squashed with https://reviews.facebook.net/D40815

Test Plan: MTR

Reviewers: hermanlee4, yoshinorim

Subscribers: