Remove expensive strcmp() calls when updating stats
Summary:
Currently three strcmp() are called to determine if the specified table is in a system database.  This is very expensive since this is done each time a row is updated.  This diff caches the information about whether the table is in a system database in the TABLE_SHARE structure so the strcmp() calls can be avoided.

https://github.com/facebook/mysql-5.6/issues/151

Squash with: https://reviews.facebook.net/D49719

Test Plan: MTR

Reviewers: hermanlee4, yoshinorim

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D53547