Persist stats after compaction is disabled
Summary:
There is currently a loophole where a regular shutdown does not persist
the cached version of table stats correctly. If the background thread
terminates before rocksdb stops all background processing, there could
be stats collected and not recorded. Note these stats are not the stats
recorded in the sst file, but rather a cached version of the sum of the
sst stats.

Terminating the background thread after all rocksdb background threads
are closed will persist those stats, but it requires the WAL to be
enabled.

Test Plan: run unit tests

Reviewers: spetrunia, yoshinorim, jkedgar

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D53697