Fix crash in table statistics
Summary:
There is a lot of confusion between the filename and system encoding
of database and table names. Generally, the storage engine uses the filename
encoding while the SQL layer uses the system encoding. However, when passing
the names back to SQL when calling the callback for table stats, we neglect to
do the conversion.

For InnoDB, the encoding is corrected to the system encoding when creating
`fil_space_t`. This should be fine, since it appears to be an in-memory
structure only, and the `table_name` and `db_name` fields were added for the
sole purpose of implementing table stats, so it should not affect anything
other than the display name in information_schema.

For RocksDB, the conversion is done right before the callback into the table
statistics update function. Also, fix a buffer overflow where filename-encoded
names were given but not enough space was allocated.

Closes #141.

Reviewed By: jkedgar

Differential Revision: https://reviews.facebook.net/D54003

fbshipit-source-id: a058e26d2ba