Fix apparent deadlock in MyRocks information schema
Summary:
When getting the column family manager through `rocksdb_get_cf_manager`, we are assigning the reference we get back to a variable on the stack. This makes a copy of the mutex inside it through the default assignment operator.

This can cause hangs because when another thread releases the mutex, the copy does not see this, and will wait forever on that mutex.

Fix this by assigning the return value to a reference.

Reviewed By: gunnarku

Differential Revision: https://reviews.facebook.net/D53979

fbshipit-source-id: d32aba7e556