Do not attempt to convert InnoDB temporary table names
Summary:
A previous diff assumed that all names within InnoDB are filename encoded and attempted to convert them. This is not true for temporary tables created by InnoDB itself. It starts with '#sql-ib', where the '#' is causing errors during the conversion.

Fix this by checking if this is a temporary table name before attempting to convert. Otherwise, just copy the name verbatim.

Squash with: D53787

Test Plan:
This code path seems to be hit by rpl.rpl_gtid_stress_crash, which is where I observed errors in the error log.

  mysqltest.sh --big rpl.rpl_gtid_stress_crash

Reviewers: jkedgar

Subscribers: