Command to purge gtid sets for given uuids
Summary:
A new command 'purge gtid' is provided to remove old gtid sets for given uuids.
The command is logged to the binlog so that slaves will also purge their gtid states
making all the replicas consistent with master's state. The command takes multiple
uuids in the following form and can only be executed by super user.

  purge gtid '551dab6a-d12b-11e5-a3ea-0002c9d31542,551dab6a-d12b-11e5-a3ea-0002c9d31543'

Purging gtid sets generated by current server is not allowed since that will break slave's
auto-positioning. This command succeeds even if a given uuid is not present in the gtid set.
If writing to binlog fails, the client will receive an error. In these cases the gtid will
get purged regardless of success of binlog write. Since this operation is idempotent, this
behavior is safe and the client running the command can retry.

Test Plan: mtr test

Reviewers: tengli, jtolmer

Reviewed By: jtolmer

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D54171
Differential Revision: https://reviews.facebook.net/D54735