Enable core dump for mysqld when starting by root then switching to non-root user later
Summary: MySQL has a bug that when starting by root then switching to a non-root user by calling setuid(), core won't be able to dumped when a core is supposed to be generated upon delivery of a signal whose default behavior is to produce a core dump. This is a old MySQL bug created in 2006 http://bugs.mysql.com/bug.php?id=21723, a fix was delivered for 5.1 in 2007 but it was wrong, the prctl(PR_SET_DUMPABLE, 1) call to enable core dump should have been called after setuid() but the fix did it before that. Yoshinori pointed this out but the issue hasn't been fixed. This diff moved the prctl(PR_SET_DUMPABLE, 1) call after the setuid() and it fixes the bug. Also mysqld option 'core-file' needs to be set to enable core dump.

Test Plan:
On dev and Jenkins servers there are no user switch when starting mysqld so the core always can be dumped. The scenario described in summary only happens in prod so manual test was done on a prod host as below (without this fix the core won't be dumped due to the bug):
(1) Build RPM with this fix and install it on a reserved prod host
(2) Add 'core-file' into the '[mysqld]' section to enable core dump
(3) Set 'innodb_buffer_pool_size' to 2G to avoid too-large core file
(4) Start mysqld and find out the PID of mysqld
(5) Use 'kill -s SIGSEGV <PID>' to kill mysqld and generate core file
(6) Use 'cat /proc/sys/kernel/core_pattern' to find out the core file pattern and location
(7) Find the core file at the location with the expected name mysqld.<PID>, the size was about 2.9GB

Reviewers: yoshinorim, jtolmer

Reviewed By: jtolmer

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D54273