Enable core dump for mysqld when running as non-root user
Summary: MySQL has a bug that when the user switching from root to a non-root user by calling setuid(), core won't be able to dumped when a core is supposed to be generated upon delivery of a signal whose default behavior is to produce a core dump. This is a old MySQL bug created in 2006 http://bugs.mysql.com/bug.php?id=21723, a fix was delivered for 5.1 in 2007 but it was wrong, the prctl(PR_SET_DUMPABLE, 1) call to enable core dump should have been called after setuid() but the fix did it before that. Yoshinori pointed this out but the issue hasn't been fixed. This diff moved the prctl(PR_SET_DUMPABLE, 1) call after the setuid() and it fixes the bug.

Test Plan: MTR

Reviewers: jtolmer, yoshinorim

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D54273