Dump core for mysqld without large memory buffer
Summary:
Core files can be very helpful for debugging production crashes but core dump has been disabled on production due to multiple reasons. One of the reasons is that the core file can be huge due to the huge InnoDB buffer pool within the process. This diff introduces a new global sys var 'innodb_dump_core_without_large_mem_buf' with type of boolean and default value of true. When this sys var is set to true, all the large memory allocation by function os_mem_alloc_large() in InnoDB will be tracked and Linux API madvise() will be called with argument MADV_DONTDUMP so that this memory chunk won't be dumped with the core. If this sys var is set to false, madvise() will be called with argument MADV_DODUMP so that this memory chunk will be dumped with the core. os_mem_alloc_large() is used to allocate large memory chunk for InnoDB buffer pool, file data merge, row log buffer, etc. Buffer pool is the main one. With this feature, the core size will be less than 3GB when the buffer pool size is 60GB so it won't impact production.

This sys var only take effect when core dump is enabled by setting 'core-file' in my.cnf.

This diff was rewritten based on the diff provided by Hui Liu from Alibaba for 5.6.16 in Feb 2014 (see https://bugs.mysql.com/bug.php?id=66826) with a couple of fixes:
(1) There is a race condition in the original diff, mem list and its mutex should be initialized when server starts instead of at the first call. Also the list should not be released when it is empty at runtime, instead, it should be released when server is shutting down.
(2) There is a bug in the sys var update -- the sys var was not updated
(3) This diff uses hash table instead of linked list. Linked list for the original diff was OK, but now we support resizable buffer pool with small buffer chunks (64MB) so the number of memory buffer that need to be tracked can reach 1000 even more, using hash table will be faster when buffer pool is resized at runtime.

There are other bugs that prevent mysqld from dumping cores, e.g. when mysqld switching user from root to a non-root user, etc. Those will be addressed in separated diffs.

Test Plan: MTR

Reviewers: jtolmer, santoshb

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D54555