Fix errors preventing clean compilation with NDEBUG defined
Summary:
The intent of these changes is to make sure that we'll compile cleanly with `NDEBUG` defined. Clean compilation is necessary for clearly separating and producing release builds. `NDEBUG` macro is //de facto// way to distinguish between debug and release code paths. It is also notably used in `<assert.h>` to determine the behavior of assertions.

There are two types of problems fixed in this set of changes:

  # Incorrect use of assertions, //i.e.//, assertions have side-effects and are evaluating invariants which are actually meant to be executed in release builds as well.
  # Warnings caused by unused variables (which when treated as errors will cause build breaks) which are fixed by adding `__attribute__((__unused__))` to the variable definitions.

Test Plan:
  ```
  mysqlbuild.sh
  mysqltest.sh
  mysqltest.sh --testset=RocksDB

  mysqlbuild.sh --release
  mysqltest.sh --release
  mysqltest.sh --release --testset=RocksDB
  ```

Reviewers: jtolmer, tianx

Reviewed By: tianx

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D54111