Change the ac_map to be a map of shared_ptr instead of unique_ptr
Summary:
The original ac_map was a map of unique_ptr<Ac_info>.  When accessing the value we had to use a reference to the value.  This diff changes that so that we use a shared_ptr<Ac_info> instead so we can just pass it around have have reference counting automatically.

Squash with: https://reviews.facebook.net/D52191

Test Plan: MTR

Reviewers: santoshb

Reviewed By: santoshb

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D54249