Fix problem with active transaction after alter with expand_fast_index_creation
Summary:
When expand_fast_index_creation is on and an alter table occurs we drop any secondary indexes and then re-add them.  These operations were occurring outside of the data copy's transaction and the restore (re-add) was causing a transaction to be started and left active.  This caused a problem after an alter that changed a table from using the MEMORY storage engine to the INNODB storage engine where an assert was hit.

https://github.com/facebook/mysql-5.6/issues/170

In the test change - one of the outputs changed from 1 to 2.  I investigated and found that with my changes the 'restoring secondary keys' state moved from after a transaction committing state to before it.  The code that did the transaction committing state switches back to the previous state when it is done.  Since the previous state is now 'restoring secondary keys' profiling shows this state twice instead of once.  I judged this to not be a concern.

Squash with: https://reviews.facebook.net/D50715

Test Plan: MTR

Reviewers: hermanlee4, jtolmer

Subscribers: