Fix a triggered assert when using expand_fast_index_creation
Summary:
When turning on expand_fast_index_creation an assert was being triggered in some code added for resizable buffer pools in InnoDB when altering a table.  The original patch that got from Oracle had the asserts, but the current Oracle code is quite different and the asserts are not obviously present.  This change just removes the questionable asserts.

https://github.com/facebook/mysql-5.6/issues/170

Squash with: https://reviews.facebook.net/D50715

Test Plan: MTR

Reviewers: hermanlee4, jtolmer, pengt

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D54447