make estimate on data_len for show table status when there is only one SST
Summary:
The Data_length and Avg_row_length are trailing statistics, meaning
they don't get updated for the current SST until the next SST is
written.  So, if rocksdb reports the data_length as 0 but there are
records, we make a reasoned estimate for the data_file_length for the
current SST.

Currently the estimate is not very good (in my opinion).  It goes through each
field in the table and gets its pack_length, and multiplies that by the # of
records.  This won't be that accurate because the actual SST has a whole bunch
of other metadata as described here:
https://github.com/facebook/rocksdb/wiki/Rocksdb-BlockBasedTable-Format

Any ideas on how we can make it better?

Test Plan: Updated rocksdb_flush_memtable_on_analyze_basic test

Reviewers: hermanlee4, yoshinorim, jkedgar

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D54489