Use fail safe RocksDB options
Summary:
The intent is to fix the issue described in https://github.com/facebook/mysql-5.6/issues/115 ("Use fail safe RocksDB options"). By utilizing the combination of `LoadLatestOptions` and `CheckOptionsCompatibility` we'll attempt to determine if opening the database is safe or it could result in a loss of data. If opening the database isn't considered to be safe then `mysqld` will refuse to start "By Design". This commit also updates the RocksDB submodule to a newer version because we need a fix from RocksDB to correctly implement this feature.

update-submodule: rocksdb

Reviewed By: jkedgar

Differential Revision: https://reviews.facebook.net/D54105

fbshipit-source-id: 67ff419f20e