Relax restrictions on replicate_same_server_id with gtid_mode
Summary:
MySQL refuses to start if replicate_same_server_id is used along
with log_slave_updates. The reason is to avoid possible infinite loops on
the slave with circular replication topology. With GTID this is not possible
as SQL thread filters out duplicate transactions.

If there is a conflict with upstream changes as part of
http://bugs.mysql.com/bug.php?id=80291, remove this diff during rebase.

Reference patch: https://github.com/facebook/mysql-5.6/commit/9847a78

Reviewed By: abhinav04sharma

Differential Revision: D9982509

fbshipit-source-id: dd7ad4ad4a8