Relax restrictions on replicate_same_server_id with gtid_mode
Summary:
MySQL refuses to start if replicate_same_server_id is used along
with log_slave_updates. The reason is to avoid possible infinite loops on
the slave with circular replication toplogy. With GTID this is not a
concern as SQL thread filters out duplicate transactions.

If there is a conflict with upstream changes as part of
http://bugs.mysql.com/bug.php?id=80291, remove this diff during rebase.

Test Plan: mtr

Reviewers: jtolmer

Subscribers: