added new hooks to watch for commits with no reviews and commits to the wrong branch
Summary:
The following script is meant to be symlinked into the .git/hooks directory.
Once it is there, it will monitor commit messages to ensure that they contain
the "Reviewed By" string, as well as check to ensure that all pushes are going
to the proper webscalesql branch.

Fixes t9689295

Test Plan:
Tested by using one of my personal github repositories to push/commit and
ensure the script works as expected.  Am relatively new to bash so any
suggestions to improve the script would be welcome!

Reviewers: jtolmer, hermanlee4

Subscribers: