Marking transaction as read only on ROLLBACK TO SAVEPOINT (#191)
Summary:
ROLLBACK TO SAVEPOINT in MyRocks returns an error if there were
modifications within the transaction, but if executing COMMIT after
that, the modifications were committed in MyRocks but excluded from
binary logs, which caused data inconsistency between master and slaves.
This diff makes transaction as rollback_only in that case, so that no
COMMIT is allowed.

squash with: https://reviews.facebook.net/D53055

Test Plan: mtr

Reviewers: jkedgar, hermanlee4, spetrunia

Subscribers: