Initialization memory in RDBSE_KEYDEF's stats structure during setup
Summary:
The stats structure for tracking cardinality allocates the memory needed
for the internal vector during the first merge() operation of the old
stats with the new stats. However, since this structure is already
shared with other thread, this reallocation can sometimes cause other
threads to segfault when reading memory. Prevent this by initializing
the vector during setup, instead of during the merge() operation.

Test Plan:
Run on the test server with the mysqlslap test that triggered the error.
Running this test with the mtr servers did not trigger a problem
unfortunately.

Reviewers: spetrunia, yoshinorim, jkedgar

Subscribers: