Fix flaky main.user_stats_rows test
Summary: The main.user_stats_rows test is pretty simple, but had one DELETE command using a join that had multiple paths the optimizer could choose.  Mostly it chose accessing table userstats first and then userstats2, but if it ever reversed this, the test would fail as a fewer number of rows get read.  I added an ANALYZE TABLE command right before this so that the optimizer would be working with specific values for the number of rows per table so that it would pick the short path and thus not fail the test because of estimates.

Test Plan: MTR (slammed user_stats_rows hundreds of times)

Reviewers: hermanlee4

Subscribers: